feat: Rename controller to spore-controller
This rename is part of the planned DRA support, as it will require to
have more than one controller, the current controller name needs to be
more descriptive.

The term spore is used here as a mostly random word, making it easy to
move to the final wording for its concept by simply replacing all
instances of the word.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>